Separate DropInViewModel and the fragment specific ViewModels.
